+++ android/app/src/main/java/com/gitcommit90/onehelm/mobile/MainActivity.java
@@ -1,0 +1,5 @@
+package com.gitcommit90.onehelm.mobile;
+
+import com.getcapacitor.BridgeActivity;
+
+public class MainActivity extends BridgeActivity {}